+++ packages/transaction-pay-controller/src/strategy/across/AcrossStrategy.ts
@@ -31,7 +31,10 @@ export class AcrossStrategy implements PayStrategy<AcrossQuote> {
       return false;
     }
 
-    if (request.transaction?.type === TransactionType.perpsDeposit) {
+    if (
+      request.transaction?.type === TransactionType.perpsDeposit ||
+      request.transaction?.type === TransactionType.perpsDepositAndOrder
+    ) {
       const supportsPerpsDeposit = actionableRequests.every((singleRequest) =>
         isSupportedAcrossPerpsDepositRequest(
           singleRequest,